2020-11-x (Bots de Telegram pt. I)/2020-10-30 (Bots de Telegram pt. I)/requirements.txt
numpy
python-telegram-bot